+++ extensions/vscode/webviews/homeView/src/components/EvenEasierDeploy.vue
@@ -191,8 +191,8 @@
           </div>
         </template>
         <template v-else>
-          <div v-if="isPreContentRecordWithoutID && isConnectPreContentRecord">
-            Is this already deployed to a Connect server? You can
+          <div v-if="isPreContentRecordWithoutID">
+            Is this already deployed to a server? You can
             <a class="webview-link" role="button" @click="onAssociateDeployment"
               >update that previous deployment</a
             >.
@@ -483,13 +483,6 @@ const isPreContentRecordWithoutID = computed(() => {
     isPreContentRecord(home.selectedContentRecord) &&
     !isPreContentRecordWithID.value
   );
-});
-
-const isConnectPreContentRecord = computed(() => {
-  const serverType =
-    home.selectedContentRecord?.serverType || ServerType.CONNECT;
-  const productType = getProductType(serverType);
-  return isConnectProduct(productType);
 });
 
 const isDismissedContentRecord = computed(() => {