+++ extensions/vscode/webviews/homeView/src/stores/home.ts
@@ -223,10 +223,9 @@ export const useHomeStore = defineStore("home", () => {
     });
   };
 
-  watch([selectedContentRecord], () => {
-    updateSelectionIsPreContentRecordState();
-    updateSelectionIsConnectContentRecordState();
-  });
+  watch([selectedContentRecord], () =>
+    updateSelectionIsPreContentRecordState(),
+  );
 
   const updateSelectionIsPreContentRecordState = () => {
     const hostConduit = useHostConduitService();
@@ -236,19 +235,6 @@ export const useHomeStore = defineStore("home", () => {
         state: isPreContentRecord(selectedContentRecord.value)
           ? "true"
           : "false",
-      },
-    });
-  };
-
-  const updateSelectionIsConnectContentRecordState = () => {
-    const hostConduit = useHostConduitService();
-    const serverType =
-      selectedContentRecord.value?.serverType || ServerType.CONNECT;
-    const productType = getProductType(serverType);
-    hostConduit.sendMsg({
-      kind: WebviewToHostMessageType.UPDATE_SELECTION_IS_CONNECT_CONTENT_RECORD,
-      content: {
-        state: isConnectProduct(productType) ? "true" : "false",
       },
     });
   };