+++ extensions/vscode/CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue where the preferred Python path was incorrectly not found when deploying #2742
 - Fixed Python dependencies scanning not respecting the active Python session #2825
 
+### Added
+
+- Added the ability to associate a Connect Cloud deployment to another Connect Cloud
+  deployment for the same account (#3067)
+
 ## [1.21.0]
 
 ### Added